#64 - Post page clean-up
