+++ changelog.json
@@ -14,7 +14,8 @@
       { "id": 4, "title": "[Enhancement]: Added `skipPages`, `skipNavigation`, and `skipSiteDesign` flags" },
       { "id": 66, "title": "[Enhancement]: Autocomplete functionality added for commands and its arguments" },
       { "id": 71, "title": "[Enhancement]: Easier debugging with ability to attach the VSCode debugger" },
-      { "id": 70, "title": "[Enhancement]: Ability to specify to clean the `QuickLaunch` and/or `TopNavigation`" }
+      { "id": 70, "title": "[Enhancement]: Ability to specify to clean the `QuickLaunch` and/or `TopNavigation`" },
+      { "id": 64, "title": "[Enhancement]: Post clean-up of removed/untouched pages" }
     ]
   },
   {
