+++ src/helpers/index.ts
@@ -1,5 +1,8 @@
+export * from './AppInsights';
 export * from './ArgumentsHelper';
+export * from './Cleanup';
 export * from './CliCommand';
+export * from './DoctorTranspiler';
 export * from './FileHelpers';
 export * from './FolderHelpers';
 export * from './FrontMatterHelper';
@@ -8,7 +11,10 @@ export * from './ListHelpers';
 export * from './MarkdownHelper';
 export * from './NavigationHelper';
 export * from './OptionsHelper';
+export * from './PagesHelper';
 export * from './ShortcodesHelpers';
 export * from './SitesHelpers';
+export * from './TelemetryHelper';
+export * from './TempDataHelper';
 export * from './execScript';
 export * from './logger';
